+++ src/Refitter.Tests/DependencyInjection/DependencyInjectionGeneratorBranchTests.cs
@@ -1,8 +1,7 @@
 using FluentAssertions;
 using Refitter.Core;
-using TUnit.Core;
 
-namespace Refitter.Tests;
+namespace Refitter.Tests.DependencyInjection;
 
 public class DependencyInjectionGeneratorBranchTests
 {
